Don't wrap <user></user> tags around empty content.
This allows the user to force the LLM to keep thinking or to
reformat the output.

You can simulate the "wait" trick in this paper: https://t.co/IGxjTt3MUy
by adding "Wait," to the end of the last <assistant> response.

Or you can do something like this from the anthropic API examples:
[
  {
    "role": "user",
    "content": "What's the Greek name for Sun? (A) Sol (B) Helios (C) Sun"
  },
  {"role": "assistant", "content": "The best answer is ("},
]