Write generated filenames to commented out line
There is a problem that when editing generated code to fix minor
problems, the file may be given a different name each time WriteCode
is called. This attempts to solve that as follows:

The WriteCode command now adds the generated filename as a comment
on the first (or second, if there is a shebang) line of the code
in the buffer. It does not get written to the file generated.

It also honors a fileanme after a comment on the first line (or
second, if there is a shebang) when writing writing a file.