Allow file, path, or full path name as arguement to WC
If an argument is included, WriteCode/WC will try to determine
if it is a directory or full path name and will try to save the
code at that path/name. starting ~ is converted to home directory
and trailing ! can be used to force directory creation if directory
does not exist.